requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

camelot-py==0.9.0
chromadb==0.4.15
pysqlite3-binary
deta==1.2.0
langchain==0.0.313
openai==0.28.1
pandas==2.0.3
PyMuPDF==1.23.4
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.27.2
streamlit-authenticator==0.2.3
tenacity==8.2.3
streamlit_extras==0.3.4
st_pages==0.4.5
llama_index==0.8.44
llama_hub==0.0.37
opencv-python-headless
